translations api factory moved to common/api/factory dir
